chore: make beta timeline the default
