+++ i18n/en.json
@@ -597,8 +597,6 @@
   "backup_setting_subtitle": "Manage background and foreground upload settings",
   "backup_settings_subtitle": "Manage upload settings",
   "backward": "Backward",
-  "beta_sync": "Beta Sync Status",
-  "beta_sync_subtitle": "Manage the new sync system",
   "biometric_auth_enabled": "Biometric authentication enabled",
   "biometric_locked_out": "You are locked out of biometric authentication",
   "biometric_no_options": "No biometric options available",
@@ -1923,6 +1921,8 @@
   "sync_albums_manual_subtitle": "Sync all uploaded videos and photos to the selected backup albums",
   "sync_local": "Sync Local",
   "sync_remote": "Sync Remote",
+  "sync_status": "Sync Status",
+  "sync_status_subtitle": "View and manage the sync system",
   "sync_upload_album_setting_subtitle": "Create and upload your photos and videos to the selected albums on Immich",
   "tag": "Tag",
   "tag_assets": "Tag assets",
