+++ mobile/lib/routing/router.gr.dart
@@ -510,22 +510,6 @@ class BackupOptionsRoute extends PageRouteInfo<void> {
 }
 
 /// generated route for
-/// [BetaSyncSettingsPage]
-class BetaSyncSettingsRoute extends PageRouteInfo<void> {
-  const BetaSyncSettingsRoute({List<PageRouteInfo>? children})
-    : super(BetaSyncSettingsRoute.name, initialChildren: children);
-
-  static const String name = 'BetaSyncSettingsRoute';
-
-  static PageInfo page = PageInfo(
-    name,
-    builder: (data) {
-      return const BetaSyncSettingsPage();
-    },
-  );
-}
-
-/// generated route for
 /// [ChangeExperiencePage]
 class ChangeExperienceRoute extends PageRouteInfo<ChangeExperienceRouteArgs> {
   ChangeExperienceRoute({
@@ -2625,6 +2609,22 @@ class SplashScreenRoute extends PageRouteInfo<void> {
     name,
     builder: (data) {
       return const SplashScreenPage();
+    },
+  );
+}
+
+/// generated route for
+/// [SyncStatusPage]
+class SyncStatusRoute extends PageRouteInfo<void> {
+  const SyncStatusRoute({List<PageRouteInfo>? children})
+    : super(SyncStatusRoute.name, initialChildren: children);
+
+  static const String name = 'SyncStatusRoute';
+
+  static PageInfo page = PageInfo(
+    name,
+    builder: (data) {
+      return const SyncStatusPage();
     },
   );
 }
