+++ mobile/lib/services/app_settings.service.dart
@@ -46,7 +46,7 @@ enum AppSettingsEnum<T> {
   syncAlbums<bool>(StoreKey.syncAlbums, null, false),
   autoEndpointSwitching<bool>(StoreKey.autoEndpointSwitching, null, false),
   photoManagerCustomFilter<bool>(StoreKey.photoManagerCustomFilter, null, true),
-  betaTimeline<bool>(StoreKey.betaTimeline, null, false),
+  betaTimeline<bool>(StoreKey.betaTimeline, null, true),
   enableBackup<bool>(StoreKey.enableBackup, null, false),
   useCellularForUploadVideos<bool>(StoreKey.useWifiForUploadVideos, null, false),
   useCellularForUploadPhotos<bool>(StoreKey.useWifiForUploadPhotos, null, false),
